Set up the module system
Now that the code is starting to expand, I figured it'll be better that I
should set up a module hierarchy with proper imports/exports in place, even
if that means that I can no longer include the build stages in the project.

So now the code just has the ts files and the bundled (by webpack) output.

Author: pm